% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_new_data_layer_to_CSV_layers_sheet.R
\name{append_new_data_layer_to_CSV_layers}
\alias{append_new_data_layer_to_CSV_layers}
\title{An Append New Data Layer to CSV Layers Sheet Function}
\usage{
append_new_data_layer_to_CSV_layers(csv_layer, ss, sheet = "CSV Layers")
}
\arguments{
\item{csv_layer}{Specify the name of the csv layer output from previous functions.}

\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get()), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().}

\item{sheet}{Sheet to append to, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number.}
}
\description{
This function allows for the appending of a new data layer into your CSV layers sheet. Can be used in conjunction with map functions such initial_ingestion() or overwite_sheet().
}
\examples{
append_new_data_layer_to_CSV_layers()
}
\keyword{CSV}
\keyword{append}
\keyword{layers}
