% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotmap_via_reticulate.R
\name{dotmap_via_reticulate}
\alias{dotmap_via_reticulate}
\title{A Python Helper Function}
\usage{
dotmap_via_reticulate(name)
}
\arguments{
\item{name}{N/A}
}
\description{
This function helps produce  a dot map from a dataframe using reticulate.
}
\keyword{dotmap}
