% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_fill_choropleth.R
\name{fill_choropleth}
\alias{fill_choropleth}
\title{A Choropleth Filling Function}
\usage{
fill_choropleth(dataframe)
}
\arguments{
\item{dataframe}{Specify the name of the dataframe that you'd like to fill.}
}
\description{
Occasionally you'll encounter a dataset that only has a value of interest at a single point in time, but you'd like to maintain that value for the duration of the timeline (e.g., discovery of a rare earth mineral deposit). This function allows you to create a wide dataframe that will maintain the value through the entire timeline. Please note that this is not a substitute for statistical techniques or to handle missing values. For those such instances, use the interpolate_choropleth() function.
}
\examples{
\dontrun{
filled_df <- fill_choropleth(dataframe)
}
}
\keyword{dataframe}
\keyword{fill}
\keyword{wide}
