% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_ingestion_function.R
\name{initial_ingestion2}
\alias{initial_ingestion2}
\title{An Initial Ingestion Function}
\usage{
initial_ingestion2(dataframe, data_layer_name)
}
\arguments{
\item{dataframe}{Specify the name of the data frame you wish to upload.}

\item{data_layer_name}{Specify the name of the new EarthTime data layer.}
}
\description{
This function allows you to create a new EarthTime data layer in Google Sheets.
}
\examples{
initial_ingestion()
}
\keyword{Google}
\keyword{ingestion}
\keyword{sheets}
\keyword{upload}
