% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_EarthTime_csv_layers.R
\name{set_EARTHTIME_CSV_LAYERS}
\alias{set_EARTHTIME_CSV_LAYERS}
\title{An Install the Default CSV Layers EarthTime Sheet in Your \code{.Renviron} File for Repeated Use Function}
\usage{
set_EARTHTIME_CSV_LAYERS(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{spreadsheet_id}{Specify the name of the csv layer output from previous functions.}

\item{GID}{Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get()), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().}

\item{sheet_name}{Sheet to append to, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number.}
}
\description{
This function will add your EarthTime CSV Layers sheet to your \code{.Renviron} file so it can be called securely without being stored in your code. After you have installed your sheet, it can be called any time by typing \code{Sys.getenv("EARTHTIME_CSV_LAYERS")} and can be used in package functions by simply typing set_EARTHTIME_CSV_LAYERS() If you do not have an \code{.Renviron} file, the function will create on for you. If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your original file for disaster recovery purposes.
}
\keyword{authorize}
