% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_EarthTime_csv_layers.R
\name{set_EarthTime_Layers_Sheet}
\alias{set_EarthTime_Layers_Sheet}
\title{An Install the Default Layers EarthTime Sheet in Your \code{.Renviron} File for Repeated Use Function}
\usage{
set_EarthTime_Layers_Sheet(
  spreadsheetId,
  GID,
  overwrite = FALSE,
  install = FALSE
)
}
\arguments{
\item{spreadsheetId}{Specify the name of the csv layer output from previous functions.}

\item{GID}{Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get()), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing \code{EarthTime_Layers_Sheet} that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions. Defaults to FALSE.}
}
\description{
This function will add your EarthTime Layers sheet to your \code{.Renviron} file so it can be called securely without being stored in your code. After you have installed your sheet, it can be called any time by typing \code{Sys.getenv("EarthTime_Layers_Sheet")} and can be used in package functions by simply typing set_EarthTime_csv_layers() If you do not have an \code{.Renviron} file, the function will create on for you. If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your original file for disaster recovery purposes.
}
\examples{
\dontrun{
if (FALSE) {
set_EarthTime_Layers_Sheet("xxxxxxxxxx-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-xxx", "0123456789", install = TRUE)
#First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
#You can check it with:
Sys.getenv("EarthTime_Layers_Sheet")
}
}
\dontrun{
if (FALSE) {
# If you need to overwrite an existing key:
set_EarthTime_Layers_Sheet("xxxxxxxxxx-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-xxx", "0123456789", install = TRUE)
# First time, relead your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("EarthTime_Layers_Sheet")
}
}
}
\keyword{authorize}
\keyword{install}
\keyword{layers}
