% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialMapTransform.R
\name{specialMapTransform}
\alias{specialMapTransform}
\title{A Long Transform for Special Map Type Function}
\usage{
specialMapTransform(
  dataframe,
  dotMap_latitude = NULL,
  dotMap_longitude = NULL,
  dotMap_dateName = NULL,
  flowMap_origin = NULL,
  flowMap_destination = NULL,
  flowMap_valueName = NULL,
  flowMap_valueColumns = NULL
)
}
\arguments{
\item{dataframe}{Specify the name of the dataframe that you'd like to transform.}

\item{dotMap_latitude}{Specify the name (in quotes) of the latitude column. You can always specify this again while invoking the create_dotMap() function. This parameter is best used when dealing with a normal "wide" style dataframe in instances where both the location values as columns.}

\item{dotMap_longitude}{Specify the name (in quotes) of the longitude column. You can always specify this again while invoking the create_dotMap() function. This parameter is best used when dealing with a normal "wide" style dataframe in instances where  both the location values as columns..}

\item{dotMap_dateName}{Specify the name (in quotes) of what you would like to call the date column in the transformed dataframe.}

\item{flowMap_origin}{Specify the name (in quotes) of the origin location. You can always specify this again while invoking the create_dotMap() function. This parameter is best used when dealing with a matrix style dataframe in instances where one location value is a single column (down) and the other location value are the column headers (across).}

\item{flowMap_destination}{Specify the name (in quotes) of the final destination. You can always specify this again while invoking the create_flowMap() function. This parameter is best used when dealing with a matrix style dataframe in instances where one location value is a single column (down) and the other location value are the column headers (across).}

\item{flowMap_valueName}{Specify the name (in quotes) of what you would like to call the value column in the transformed dataframe.}

\item{value_columns}{A concatenated vector of the pre-transformed dataframe's date columns. Use, e.g., value_columns = c(3:223)}
}
\description{
While most cases required a choropleth or bubble style "wide" dataframe for ingestion into EarthTime, a few special map types in EarthTime (i.e. - line strings, point flows, points, and raster maps) need to be formatted as "long" dataframes types before using create_dotMap() and/or create_flowMap() to generate a .bin  file. This function transforms the dataframe in an appropriate format to successfully utilize the aforementioned functions. Use the dotMap_... and flowMap_... parameters to transform the data for Dot Maps and Flow Maps, respectively.
}
\examples{
\dontrun{
long_df_dotMap <- specialMapTransform(dataframe, dotMap_latitude, dotMap_longitude, dotMap_dateName)
}
\dontrun{
long_df_flowMap <- specialMapTransform(dataframe, flowMap_origin, flowMap_destination, flowMap_valueName, flowMap_valueColumns)
}
}
\keyword{dataframe}
\keyword{dot}
\keyword{flow}
\keyword{long}
\keyword{map}
\keyword{special}
\keyword{transform}
