% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_categories_function.R
\name{split_categories}
\alias{split_categories}
\title{A Splitting Categories Function}
\usage{
split_categories(dataframe, categories_column)
}
\arguments{
\item{dataframe}{Specify the name of the wide dataframe that should be split.}

\item{categories_column}{Specify the column name (in quotes) or the column number that contains the various categories.}
}
\description{
This function splits a single dataframe into multiple dataframes when there is more than one category present that would be more beneficial as individual data layers in EarthTime. It assumes that the dataframe is already in a "wide" format. If not, please use the choropleth() function in this package before invoking split_categories().
}
\examples{
\dontrun{
split_categories(dataframe, categories_column)
}
}
\keyword{categories}
\keyword{split}
