% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_function.R
\name{choropleth}
\alias{choropleth}
\title{A Choropleth Function}
\usage{
choropleth(dataframe, date, value)
}
\arguments{
\item{dataframe}{Specify the name of the long data frame you wish to format.}

\item{date}{Specify the name of the date column.}

\item{value}{Specify the name of the value you'd like to spread across the wide data frame.}
}
\description{
This function allows for the transformation of a long dataframe into a wide dataframe for an EarthTime choropleth.
}
\examples{
choropleth()
}
\keyword{choropleth}
