% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flowMap_function.R
\name{create_flowMap}
\alias{create_flowMap}
\title{A Flow Map Creation Function}
\usage{
create_flowMap(
  dataframe,
  origin,
  destination,
  value_column,
  date_column,
  projected_coordinate_system = "epsg_3857",
  centroids_geojson = "country-centroids.geojson",
  centroids_json = "centroids_by_iso_alpha_3.json",
  output_dir = getwd()
)
}
\arguments{
\item{dataframe}{Specify the name of the dataframe that you'd like to transform into a dot map .json file.}

\item{origin}{Specify the column name (in quotes) that contains the latitudinal coordinates.}

\item{destination}{Specify the column name (in quotes) that contains the longitudinal coordinates.}

\item{value_column}{Specify the column name (in quotes) that contains the primary value being visualized.}

\item{date_column}{Specify the column name (in quotes) that contains the date category.}

\item{projected_coordinate_system}{A flat, two-dimensional representation of the Earth based on a sphere or spheroid geographic coordinate system, but it uses linear units of measure for coordinates, so that calculations of distance and area are easily done in terms of those same units. Defaults to EPSG:3857 (WGS 84 / Pseudo-Mercator) used for rendering maps in Spherical Mercator, Google Maps, OpenStreetMap, Bing, ArcGIS, ESRI, etc.}

\item{centroids_geojson}{Defaults to "country-centroids.geojson" included with the package.}

\item{centroids_json}{Defaults to "centroids_by_iso_alpha_3.json" included with the package.}

\item{output_dir}{Output director for .json file. Defaults to current working directory.}
}
\description{
This function creates a flow map from a dataframe. It will output a .json file in your working directory for you to later push to a specific EarthTime instance.
}
\examples{
\dontrun{
create_flowMap <-  function(dataframe, origin, destination, value_column, date_column)
}
}
\keyword{flow}
\keyword{map}
