% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_csv_layer_row_function.R
\name{generate_csv_layer_row}
\alias{generate_csv_layer_row}
\title{A Generating Row for CSV Layer Function}
\usage{
generate_csv_layer_row(
  Enabled = TRUE,
  Share_Link_Identifier,
  Category,
  Name,
  Credits,
  Base_Layer = NA,
  Custom_slider_ticks = NA,
  Start_Date,
  End_Date,
  Step = 1,
  URL,
  Scaling,
  Color_Scaling = NA,
  Map_Type,
  Color,
  External_GeoJSON = NA,
  Name_Key = NA,
  Show_Graph = NA,
  Graph_Title = NA,
  Graph_Plot_Colors = NA,
  Graph_Plots_First_Visible = NA,
  Graph_X_Axis_Label = NA,
  Graph_Y_Axis_Label = NA,
  Graph_Y_Axis_Min = NA,
  Graph_Y_Axis_Max = NA,
  Graph_X_Axis_Label_Interval = NA,
  Legend_Content = "auto",
  Legend_Key,
  Load_Data_Function = NA,
  Set_Data_Function = NA,
  Number_of_Levels = NA,
  Number_of_Attributes = NA,
  Vertex_Shader = NA,
  Fragment_Shader = NA,
  Draw_Function = NA,
  Playback_Rate = 1,
  Master_Playback_Rate = 1,
  Colormap_Src,
  Layer_Description = NA,
  Featured_Themes = NA,
  Draw_Options = NA,
  Set_Data_Options = NA,
  Extras_Options = NA,
  Draw_Order = NA,
  Timeline_Type = NA,
  Layer_Constraints = NA,
  Mapbox = NA,
  Draw_Layer_Function = NA
)
}
\arguments{
\item{Enabled}{Determines if the data layers appears in the Explore view of EarthTime. Defaults to TRUE.}

\item{Share_Link_Identifier}{This is the unique identifier that will be issued to the data layer. Please ensure that there are no other data layers with the same name; otherwise, it will cause conflicts and normal functioning and operation of the EarthTime will be impacted.}

\item{Category}{Name of the data layer's category}

\item{Name}{Name of the data layer. Defaults to the dataframe's name}

\item{Credits}{Citation for the data layers}

\item{Base_Layer}{Ties a specific base layer with the data layer. Defaults to NULL.}

\item{Custom_slider_ticks}{CREATE LAB special use case. Defaults to NULL.}

\item{Start_Date}{Column number for the first date in the data frame.}

\item{End_Date}{Column number for the last date in the data frame.}

\item{Step}{Time period interval to run the data in the timeline. Defaults to 1.}

\item{URL}{Proxy URL for data layer to be read into the EarthTime engine.}

\item{Scaling}{d3-scale variables. Scales are a convenient abstraction for a fundamental task in visualization: mapping a dimension of abstract data to a visual representation. Although most often used for position-encoding quantitative data, such as mapping a measurement in meters to a position in pixels for dots or color spectrum in EarthTime; moreover, these scales are meant to represent the quantitative data ingested into EarthTime. Depending on the distribution, one should consider the following scales: "Linear", "Power", "Log", or "SquareRoot". Inputting the aforementioned scale variables will output a domain of "minValue, maxValue" with a range of "O,100" for the d3 function.}

\item{Color_Scaling}{CREATE LAB special use case. Defaults to NULL.}

\item{Map_Type}{Type of map to visualize. Acceptable variables include: "Choropleth", "Bubble", "Bivalent Bubble", "Point", "Point Flow"}

\item{Color}{RGB color value. Acceptable format is "0.0, 0.0, 0.0"}

\item{External_GeoJSON}{Link to external GeoJSON file. Defaults to NULL.}

\item{Name_Key}{Data column name for GEOJSON shapes. Defaults to NULL.}

\item{Show_Graph}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Title}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Plot_Colors}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Plots_First_Visible}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_X_Axis_Label}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Y_Axis_Label}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Y_Axis_Min}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_Y_Axis_Max}{CREATE LAB special use case. Defaults to NULL.}

\item{Graph_X_Axis_Label_Interval}{CREATE LAB special use case. Defaults to NULL.}

\item{Legend_Content}{Legend content displayed in explore view. Defaults to "auto". For other legend types, input HTML here.}

\item{Legend_Key}{Short description of the data layer for use in the legend.}

\item{Load_Data_Function}{For use in point map types. Defaults to NULL; otherwise, input function such as: "WebGLVectorTile2.prototype._loadData".}

\item{Set_Data_Function}{For use in point map types. Defaults to NULL; otherwise, input function such as: "WebGLVectorTile2.prototype._setBufferData".}

\item{Number_of_Levels}{CREATE LAB special use case. Defaults to NULL.}

\item{Number_of_Attributes}{For use in point map types. Defaults to NULL; otherwise, input number of attributes such as "5".}

\item{Vertex_Shader}{For use in point map types. Defaults to NULL; otherwise, input shader such as: "WebGLVectorTile2.PointSizeColorEpochVertexShader".}

\item{Fragment_Shader}{For use in point map types. Defaults to NULL; otherwise, input shader such as: "WebGLVectorTile2.PointColorFragmentShader".}

\item{Draw_Function}{For use in point map types. Defaults to NULL; otherwise, input function such as: "WebGLVectorTile2.prototype._drawPointSizeColorEpoch".}

\item{Playback_Rate}{Natural speed of the data layer timeline. Defaults to 1.}

\item{Master_Playback_Rate}{Override speed of the data layer timeline. Defaults to 1.}

\item{Colormap_Src}{URL of the color spectrum for choropleths. Refer to "https://tiles.earthtime.org/colormaps/black-background.html".}

\item{Layer_Description}{CREATE LAB special use case. Defaults to NULL.}

\item{Featured_Themes}{CREATE LAB special use case. Defaults to NULL.}

\item{Draw_Options}{CREATE LAB special use case. Defaults to NULL.}

\item{Set_Data_Options}{CREATE LAB special use case. Defaults to NULL.}

\item{Extras_Options}{CREATE LAB special use case. Defaults to NULL.}

\item{Draw_Order}{CREATE LAB special use case. Defaults to NULL.}

\item{Timeline_Type}{CREATE LAB special use case. Defaults to NULL.}

\item{Layer_Constraints}{CREATE LAB special use case. Defaults to NULL.}

\item{Mapbox}{CREATE LAB special use case. Defaults to NULL.}

\item{Draw_Layer_Function}{CREATE LAB special use case. Defaults to NULL.}
}
\description{
Generally used after the overwrite_sheet_function(), the generate_csv_layer_row() function allows for the generation of a new data layer for input into the EarthTime CSV layers sheet. This function is usually followed up by append_new_data_layer_to_CSV_layers().
}
\examples{
append_new_data_layer_to_CSV_layers()
}
\keyword{CSV}
\keyword{append}
\keyword{layers}
