% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_sheet_function.R
\name{overwrite_sheet}
\alias{overwrite_sheet}
\title{An Overwrite Sheet Function}
\usage{
overwrite_sheet(dataframe, spreadsheetId, sheetId)
}
\arguments{
\item{dataframe}{Specify the name of data frame you wish to upload.}

\item{spreadsheetId}{Specify the spreadsheet ID (workbook ID) of your Google sheet.}

\item{sheetId}{Specify the sheet that you would like to overwrite. "sheetId" is lso referred to as Google Grid ID, or "gid".}

\item{sheet_name}{If there is a different sheet name other than "Sheet1", specify it here; otherwise, it will automatically default to "Sheet1".}
}
\description{
This function allows you to upload and overwrite an EarthTime data layer in Google Sheets.
}
\examples{
overwrite_sheet()
}
\keyword{Google}
\keyword{overwrite}
\keyword{sheets}
\keyword{upload}
